pyproject.toml
[project]
name = "eduremus-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
target_version = ["py312", "py313"]
dependencies = [
    "django==5.2.4",
    "django-jazzmin==3.0.1",
    "django-split-settings==1.3.2",
    "faker==37.4.2",
]

[dependency-groups]
dev = [
    "black==25.1.0",
    "django-extensions==4.1",
    "ipython==9.4.0",
    "werkzeug==3.1.3",
]
